+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 9.1
 - Fix stored XSS security issue: decode HTML entities from URL in PreparePHP_SELF.fnc.php, thanks to @domiee13
 - Capitalize month when date is only month and year in Dashboard.inc.php
 - Add decimal & thousands separator configuration in Help_en.php, Currency.fnc.php, Configuration.php, rosariosis.sql & rosariosis_fr.sql
+- Use Currency() for Food Service Balance value in Widget.php & StaffWidget.php
 
 Changes in 9.0
 --------------